+++ packages/components/input/src/index.ts
@@ -1,10 +1,12 @@
 import Input from "./input";
+import Textarea from "./textarea";
 
 // export types
 export type {InputProps} from "./input";
+export type {TextAreaProps} from "./textarea";
 
 // export hooks
 export {useInput} from "./use-input";
 
 // export component
-export {Input};
+export {Input, Textarea};