+++ packages/components/radio/src/radio-group.tsx
@@ -6,8 +6,15 @@ import {UseRadioGroupProps, useRadioGroup} from "./use-radio-group";
 export interface RadioGroupProps extends Omit<UseRadioGroupProps, "ref"> {}
 
 const RadioGroup = forwardRef<RadioGroupProps, "div">((props, ref) => {
-  const {Component, children, label, context, getGroupProps, getLabelProps, getWrapperProps} =
-    useRadioGroup({ref, ...props});
+  const {
+    Component,
+    children,
+    label,
+    context,
+    getGroupProps,
+    getLabelProps,
+    getWrapperProps,
+  } = useRadioGroup({ref, ...props});
 
   return (
     <Component {...getGroupProps()}>