+++ packages/hooks/use-ssr/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @nextui-org/use-ssr
 
+## 0.0.0-dev-v2-20230405205125
+
+### Patch Changes
+
+- Native events exposed in all form elements
+
+## 0.0.0-dev-v2-20230405200614
+
+### Patch Changes
+
+- Textarea component added
+
 ## 0.0.0-dev-v2-20230405030905
 
 ### Patch Changes